pyproject.toml
[project]
name = "paws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.59.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
